Move the gui and coordinate system transformation sections from the master branch to the rtthread branch

+++ program/device/sensor_decode.h
@@ -32,7 +32,7 @@
  * */
 typedef struct {
     short north;
-    short kalman_north;
+    float kalman_north;
     short ax;
     short ay;
     short az;
